Adapt: createDeployment was always destroying deployment Don't do that. Only destroy on dryRun.

+++ adapt/src/ops/createDeployment.ts
@@ -42,6 +42,6 @@ export async function createDeployment(options: CreateOptions): Promise<DeploySt
             ...buildOpts
         });
     } finally {
-        await destroyDeployment(server, deployment.deployID);
+        if (finalOptions.dryRun) await destroyDeployment(server, deployment.deployID);
     }
 }
